requirements.txt
opencv-python-headless
numpy
torch
torchvision
torchaudio
ultralytics
pillow
transformers
scipy
